Quantum_Hackathon/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
pillow==11.1.0
pennylane==0.32.0
numpy==2.2.4
torchviz>=0.0.2
graphviz>=0.20.1 